Change: split Xor8 into a builder and filter
`Xor8Builder` is in charge of building a filter but does not provide
query APIs.

`Xor8` now becomes a pure filter implementation that just provides query
APIs.

By splitting these two roles the data structures are simplified, e.g. no
need to unpack an `Option` of the `keys`.
