+++ src/xor8/mod.rs
@@ -1,0 +1,7 @@
+mod builder;
+mod filter;
+#[cfg(test)]
+mod tests;
+
+pub use builder::Xor8Builder;
+pub use filter::Xor8;
